Delete unused keccak circuit; rename keccak2 to keccak
