+++ zirgen/circuit/keccak/cpp/trace.h
@@ -1,4 +1,4 @@
-// Copyright 2024 RISC Zero, Inc.
+// Copyright 2025 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@
 #include "risc0/fp/fp.h"
 #include "risc0/fp/fpext.h"
 
-namespace zirgen::keccak2 {
+namespace zirgen::keccak {
 
 using Fp = risc0::Fp;
 using FpExt = risc0::FpExt;
@@ -72,4 +72,4 @@ struct ExecutionTrace {
   GlobalTraceGroup global;
 };
 
-} // namespace zirgen::keccak2
+} // namespace zirgen::keccak
