+++ zirgen/circuit/keccak/is_zero.zir
