--- zirgen/circuit/keccak/test0k.nop
@@ -1,11 +1,0 @@
-// RUN: zirgen -I %S --test %s --test-cycles 3000 --input-data-hex 010000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080c5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a4700000000000000000 2>&1 | FileCheck %s
-
-import keccak;
-
-test TestKeccak00k { // test_cycles >= 2974
-// CHECK-LABEL: Running TestKeccak00k
-// CHECK: sha256output: 0xc376,0xd218 0xa46f,0x1750 0xd163,0xea2 0xf9dc,0xb7a 0xc1ca,0xe1a6 0xa6a,0x259b 0x1023,0x7c1a 0xa863,0x32a8
-// CHECK: final accum: [0, 0, 0, 0]
-  Top ()
-}
-
