--- zirgen/circuit/keccak/test0s.nop
@@ -1,11 +1,0 @@
-// RUN: zirgen -I %S --test %s --test-cycles 3000 --input-data-hex 010000000000000006000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080a7ffc6f8bf1ed76651c14756a061d662f580ff4de43b49fa82d80a4b80f8434a0000000000000000 2>&1 | FileCheck %s
-
-import keccak;
-
-test TestKeccak00s { // test_cycles >= 2974
-// CHECK-LABEL: Running TestKeccak00s
-// CHECK: sha256output: 0xd7da,0x872c 0x1921,0x7174 0x4dcf,0x92da 0xa191,0xf653 0x4c9f,0x10be 0xac8,0x7e90 0xf316,0x6b75 0x97c5,0xcca0
-// CHECK: final accum: [0, 0, 0, 0]
-  Top ()
-}
-
