--- zirgen/circuit/keccak/test1k.zir
@@ -1,11 +1,0 @@
-// RUN: zirgen -I %S --test %s --test-cycles 3000 --input-data-hex 010000000000000054686520717569636B2062726F776E20666F78206A756D7073206F76657220746865206C617A7920646F672E0100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080578951e24efd62a3d63a86f7cd19aaa53c898fe287d2552133220370240b572d0000000000000000 2>&1 | FileCheck %s
-
-import keccak;
-
-test TestKeccak01k { // test_cycles >= 2974
-// CHECK-LABEL: Running TestKeccak01k
-// CHECK: sha256output: 0x95b3,0x6374 0x988e,0x6e0a 0xec7c,0xb317 0x54fd,0x4847 0xb009,0x9392 0xdafc,0x4759 0x3f57,0x7866 0x8a26,0xc723
-// CHECK: final accum: [0, 0, 0, 0]
-  Top ()
-}
-
