--- zirgen/circuit/keccak/test1s.nop
@@ -1,11 +1,0 @@
-// RUN: zirgen -I %S --test %s --test-cycles 3000 --input-data-hex 010000000000000054686520717569636B2062726F776E20666F78206A756D7073206F76657220746865206C617A7920646F672E0600000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080a80f839cd4f83f6c3dafc87feae470045e4eb0d366397d5c6ce34ba1739f734d0000000000000000 2>&1 | FileCheck %s
-
-import keccak;
-
-test TestKeccak01s { // test_cycles >= 2974
-// CHECK-LABEL: Running TestKeccak01s
-// CHECK: sha256output: 0xe656,0x6db2 0xbe09,0x4313 0x98b6,0xe5eb 0xdd04,0x85b1 0xdd60,0x6831 0x18a,0x5515 0xab3b,0xfffd 0xa2f3,0xdbe0
-// CHECK: final accum: [0, 0, 0, 0]
-  Top ()
-}
-
