--- zirgen/circuit/keccak/test_keccak_01.txt
@@ -1,3 +1,0 @@
-010000000000000054686520717569636B2062726F776E20666F78206A756D7073206F76657220746865206C617A7920646F672E0100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080578951e24efd62a3d63a86f7cd19aaa53c898fe287d2552133220370240b572d0000000000000000
-
-The quick brown fox jumps over the lazy dog.
